move: pluto/eris/triton -> curve
